Module Miou.Mutex

type t

The type of mutexes.

val create : unit -> t

Return a new mutex.

val unlock : t -> unit

Unlock the given mutex. Other tasks suspended trying to lock the mutex will restart. The mutex must have been previously locked by the thread that calls unlock.

raises Sys_error
was not raised when unlocking an unlocked mutex or when unlocking a mutex from a different task.

val lock : t -> unit

Lock the given mutex. Only one task can have the mutex locked at a time. A task that attempts to lock a mutex already locked by another thread will suspend until the other thread unlocks the mutex.

val try_lock : t -> bool

Same as lock, but does not suspend the calling thread if the mutex is already locked: just return false immediately in that case. If the mutex is unlocked, lock it and return true.

val protect : t -> (unit -> 'a) -> 'a

protect t fn runs fn in a critical section where t is locked (atomically); it then takes care of releasing t whether fn returned a value or raised an exception.

The unlocking operation is guaranteed to always takes place, even in the event a cancellation is ordered by the parent task.